pyproject.toml
[tool.poetry]
name = "gs-aws-to-gcp-workload-identity"
version = "0.1.0"
description = "Helper for obtaining GCP Tokens via AWS-to-GCP Workload Identity Federation"
authors = ["Brad Geesaman"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.24.40"
botocore = "^1.27.40"
requests = "^2.28.1"
jwt = "^1.3.1"
google-auth = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
